notebook_to_pdf_agent/pyproject.toml
[project]
name = "notebook-to-pdf-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.2",
    "nbconvert[webpdf]>=7.17.0",
    "ollama>=0.6.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
